Steps: Export promoted tags
