--- cmd/ci-operator-checkconfig/main_test.go
@@ -1,66 +1,0 @@
-package main
-
-import (
-	"reflect"
-	"testing"
-
-	"k8s.io/utils/diff"
-
-	"github.com/openshift/ci-tools/pkg/api"
-)
-
-func TestPromotedTags(t *testing.T) {
-	var testCases = []struct {
-		name     string
-		input    *api.ReleaseBuildConfiguration
-		expected []api.ImageStreamTagReference
-	}{
-		{
-			name:     "no promotion, no output",
-			input:    &api.ReleaseBuildConfiguration{},
-			expected: nil,
-		},
-		{
-			name: "promoted image means output tags",
-			input: &api.ReleaseBuildConfiguration{
-				Images: []api.ProjectDirectoryImageBuildStepConfiguration{
-					{To: api.PipelineImageStreamTagReference("foo")},
-				},
-				PromotionConfiguration: &api.PromotionConfiguration{
-					Namespace: "roger",
-					Name:      "fred",
-				},
-			},
-			expected: []api.ImageStreamTagReference{{
-				Namespace: "roger",
-				Name:      "fred",
-				Tag:       "foo",
-			}},
-		},
-		{
-			name: "promoted image by tag means output tags",
-			input: &api.ReleaseBuildConfiguration{
-				Images: []api.ProjectDirectoryImageBuildStepConfiguration{
-					{To: api.PipelineImageStreamTagReference("foo")},
-				},
-				PromotionConfiguration: &api.PromotionConfiguration{
-					Namespace: "roger",
-					Tag:       "fred",
-				},
-			},
-			expected: []api.ImageStreamTagReference{{
-				Namespace: "roger",
-				Name:      "foo",
-				Tag:       "fred",
-			}},
-		},
-	}
-
-	for _, testCase := range testCases {
-		t.Run(testCase.name, func(t *testing.T) {
-			if actual, expected := promotedTags(testCase.input), testCase.expected; !reflect.DeepEqual(actual, expected) {
-				t.Errorf("%s: got incorrect promoted tags: %v", testCase.name, diff.ObjectDiff(actual, expected))
-			}
-		})
-	}
-}
