+++ pkg/steps/release/promote_test.go
@@ -140,13 +140,69 @@ func TestToPromote(t *testing.T) {
 
 	for _, test := range testCases {
 		t.Run(test.name, func(t *testing.T) {
-			bySource, names := ToPromote(test.config, test.images, test.requiredImages)
+			bySource, names := toPromote(test.config, test.images, test.requiredImages)
 			if actual, expected := bySource, test.expectedBySource; !reflect.DeepEqual(actual, expected) {
 				t.Errorf("%s: got incorrect tags by source: %s", test.name, diff.ObjectDiff(actual, expected))
 			}
 
 			if actual, expected := names, test.expectedNames; !reflect.DeepEqual(actual, expected) {
 				t.Errorf("%s: got incorrect names: %s", test.name, diff.ObjectDiff(actual, expected))
+			}
+		})
+	}
+}
+
+func TestPromotedTags(t *testing.T) {
+	var testCases = []struct {
+		name     string
+		input    *api.ReleaseBuildConfiguration
+		expected []api.ImageStreamTagReference
+	}{
+		{
+			name:     "no promotion, no output",
+			input:    &api.ReleaseBuildConfiguration{},
+			expected: nil,
+		},
+		{
+			name: "promoted image means output tags",
+			input: &api.ReleaseBuildConfiguration{
+				Images: []api.ProjectDirectoryImageBuildStepConfiguration{
+					{To: api.PipelineImageStreamTagReference("foo")},
+				},
+				PromotionConfiguration: &api.PromotionConfiguration{
+					Namespace: "roger",
+					Name:      "fred",
+				},
+			},
+			expected: []api.ImageStreamTagReference{{
+				Namespace: "roger",
+				Name:      "fred",
+				Tag:       "foo",
+			}},
+		},
+		{
+			name: "promoted image by tag means output tags",
+			input: &api.ReleaseBuildConfiguration{
+				Images: []api.ProjectDirectoryImageBuildStepConfiguration{
+					{To: api.PipelineImageStreamTagReference("foo")},
+				},
+				PromotionConfiguration: &api.PromotionConfiguration{
+					Namespace: "roger",
+					Tag:       "fred",
+				},
+			},
+			expected: []api.ImageStreamTagReference{{
+				Namespace: "roger",
+				Name:      "foo",
+				Tag:       "fred",
+			}},
+		},
+	}
+
+	for _, testCase := range testCases {
+		t.Run(testCase.name, func(t *testing.T) {
+			if actual, expected := PromotedTags(testCase.input), testCase.expected; !reflect.DeepEqual(actual, expected) {
+				t.Errorf("%s: got incorrect promoted tags: %v", testCase.name, diff.ObjectDiff(actual, expected))
 			}
 		})
 	}
